api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
weaviate-client>=3.24.1
scikit-learn>=1.0.0
numpy>=1.21.0
pydantic>=2.0.0